update handler & fix some bugs

+++ qlib/contrib/backtest/position.py
@@ -160,7 +160,7 @@ class Position:
     def save_position(self, path, last_trade_date):
         path = pathlib.Path(path)
         p = copy.deepcopy(self.position)
-        cash = pd.Series()
+        cash = pd.Series(dtype=np.float)
         cash["init_cash"] = self.init_cash
         cash["cash"] = p["cash"]
         cash["today_account_value"] = p["today_account_value"]
